move workspace test to encrypted storage
